+++ CHANGELOG.md
@@ -2,6 +2,6 @@
 
 ## major
 * Add aggregate file paths from a list of match strings
-* Add XML update and remove actions
+* Add XML update, create, and remove actions
 * Add YAML configuration
 * Add logging